+++ Example/Example_BusinessComponents/MockData/RowData.swift
@@ -31,11 +31,11 @@ class RowData: NSObject, MLBusinessRowData {
     }
     
     func getMainDescription() -> [MLBusinessRowMainDescriptionData]? {
-        return [RowMainDescriptionData(type: "image", content: "https://i.ibb.co/2KcdLdy/16-px.png", color: "#454545"),
-                RowMainDescriptionData(type: "text", content: "598m", color: "#454545"),
-                RowMainDescriptionData(type: "text", content: " · ", color: "#454545"),
-                RowMainDescriptionData(type: "image", content: "https://i.ibb.co/prSV6dY/estrella-android.png", color: "#454545"),
-                RowMainDescriptionData(type: "text", content: "4.3 (24)", color: "#454545"),]
+        return [RowMainDescriptionData(type: "image", content: "https://i.ibb.co/2KcdLdy/16-px.png", color: "#737373"),
+                RowMainDescriptionData(type: "text", content: "598m", color: "#737373"),
+                RowMainDescriptionData(type: "text", content: " · ", color: "#737373"),
+                RowMainDescriptionData(type: "image", content: "https://i.ibb.co/prSV6dY/estrella-android.png", color: "#737373"),
+                RowMainDescriptionData(type: "text", content: "4.3 (24)", color: "#737373"),]
     }
     
     func getMainSecondaryDescription() -> [MLBusinessMultipleDescriptionModel]? {
@@ -77,6 +77,10 @@ class RowData: NSObject, MLBusinessRowData {
     
     func getLink() -> String? {
         return "mercadopago://scanqr"
+    }
+    
+    func getMainTitleStatus() -> String? {
+        return "CLOSED"
     }
 }