+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Sin publicar
 - Fix en `setAdditionalEdgeInsets(with:)` para el Cover Carousel
+- Added closed status support to title and subtile on Hybrid Row
 
 # v1.30.0
 🚀 1.30.0 🚀